Remove chunkoverlap task
We got feedback and eval results that showed this task wasn't working as expected.

Looking at the implementation, I am not too surprised, as I think how it is implemented will interact poorly with the core text splitting.

If we find a use case for this, it would be better to integrate chunk overlap at the level of actually generating the chunks so we can find better split points. We can also eval against other implementations as well.