test: Explicitly test new index configuration structure
The changes introduced in the previous commit were covered by a test
that randomly selects either semantic or instructable. Running the test
suite multiple times in a row is the only way to confirm with high
probability that the new structure (which now includes a variant) is
correct.

This commit splits this test out into two sub-tests which check the
semantic and instructable variants individually, with the usual
field-level randomisation.