feat: new Pharia Kernel connector
To allow for using Kernel Skills as either subtasks within an existing Task, or to pass into the Eval framework, this is a basic wrapper around a task that allows for invoking a skill as a Task.

The Pydantic types are passed into the constructor so that they can be callable and also to make mypy generics happy. You can check out the test to see how this would be used.